+++ wlts/collections/collection_manager.py
@@ -1,13 +1,14 @@
 #
 # This file is part of Web Land Trajectory Service.
-# Copyright (C) 2019 INPE.
+# Copyright (C) 2019-2020 INPE.
 #
 # Web Land Trajectory Service is free software; you can redistribute it and/or modify it
 # under the terms of the MIT License; see LICENSE file for more details.
 #
 """WLTS Collection Manager."""
-import pkg_resources
 from json import loads as json_loads
+
+import pkg_resources
 
 from wlts.collections.feature_collection import FeatureCollection
 from wlts.collections.image_collection import ImageCollection
@@ -18,15 +19,24 @@ class CollectionFactory:
 
     @staticmethod
     def make(collection_type, collections_info):
-        """Factory method creates Collection."""
+        """Factory method to creates a collection.
+
+        Args:
+            collection_type (str): The collection type to be create.
+            collections_info (dict): The collection information.
+
+        Returns:
+            collection: A collection object.
+        """
         factorys = {"feature_collection": "FeatureCollection", "image_collection": "ImageCollection"}
 
         collection = eval(factorys[collection_type])(collections_info)
 
         return collection
 
+
 class CollectionManager:
-    """CollectionManager Class."""
+    """This is a singleton to manage all collections instances available."""
 
     _collections = list()
 
@@ -34,36 +44,54 @@ class CollectionManager:
 
     def __init__(self):
         """Virtually private constructor."""
-        if CollectionManager.__instance != None:
+        if CollectionManager.__instance is not None:
             raise Exception("This class is a singleton!")
         else:
             CollectionManager.__instance = self
             CollectionManager.__instance.load_all()
 
     @staticmethod
-    def getInstance():
+    def get_instance():
         """Static access method."""
-        if CollectionManager.__instance == None:
+        if CollectionManager.__instance is None:
             CollectionManager()
         return CollectionManager.__instance
 
-    def insert(self, collectionType, collection_info):
-        """Insert Collection."""
-        collection = CollectionFactory.make(collectionType, collection_info)
+    def insert(self, collection_type, collection_info):
+        """Method to creates a new collection and stores in list of collections.
+
+        Args:
+            collection_type (str): The collection type to be create.
+            collection_info (dict): The collection information.
+        """
+        collection = CollectionFactory.make(collection_type, collection_info)
         self._collections.append(collection)
 
-
     def get_collection(self, name):
-        """Get Collection."""
+        """Return the collection.
+
+        Args:
+            name (str): Identifier (name) of an collection.
+
+        Returns:
+            collection: A collection available in the server.
+
+        Raises:
+            RuntimeError: If the collection not found.
+        """
         try:
             for collection in self._collections:
                 if name == collection.get_name():
                     return collection
-        except:
-            return None
+        except ValueError:
+            raise RuntimeError(f"Collection {name} not found!")
 
     def collection_names(self):
-        """Get Name of all collections avaliable."""
+        """Return all available collections.
+
+        Returns:
+            list: A list with all collections identifier (name) available in the server.
+        """
         collections_names = list()
 
         for collection in self._collections:
@@ -72,12 +100,13 @@ class CollectionManager:
         return collections_names
 
     def get_all_collections(self):
-        """Get all Collections."""
+        """Returns a list with all collections objects."""
         return self._collections
 
     def load_all(self):
-        """Load all Collection."""
-        json_string_feature = pkg_resources.resource_string('wlts', '/json_configs/feature_collection.json').decode('utf-8')
+        """Creates all collection based on json of image and feature collection."""
+        json_string_feature = pkg_resources.resource_string('wlts', '/json_configs/feature_collection.json').decode(
+            'utf-8')
 
         json_string_image = pkg_resources.resource_string('wlts', '/json_configs/image_collection.json').decode('utf-8')
 
@@ -93,5 +122,6 @@ class CollectionManager:
             image_collection = config_image["image_collection"]
             for img_collection in image_collection:
                 self.insert("image_collection", img_collection)
+
 
 collection_manager = CollectionManager()