+++ wlts/collections/image_collection.py
@@ -1,19 +1,24 @@
 #
 # This file is part of Web Land Trajectory Service.
-# Copyright (C) 2019 INPE.
+# Copyright (C) 2019-2020 INPE.
 #
 # Web Land Trajectory Service is free software; you can redistribute it and/or modify it
 # under the terms of the MIT License; see LICENSE file for more details.
 #
 """WLTS Image Collection Class."""
-from .collection import Collection
 from ..utils import get_date_from_str
+from .collection import Collection
+
 
 class ImageCollection(Collection):
-    """ImageCollection Class."""
+    """Implement a image collection."""
 
     def __init__(self, collections_info):
-        """Creates ImageCollection."""
+        """Creates ImageCollection.
+
+        Args:
+            collections_info (dict): The collection information.
+        """
         super().__init__(collections_info["name"],
                          collections_info["authority_name"],
                          collections_info["description"],
@@ -33,11 +38,22 @@ class ImageCollection(Collection):
         self.timeline = collections_info["timeline"]
 
     def collection_type(self):
-        """Get Collection Image Type."""
+        """Return the collection type."""
         return "Image"
 
     def trajectory(self, tj_attr, x, y, start_date, end_date):
-        """Get Trajectory."""
+        """Return the trajectory.
+
+        Args:
+            tj_attr (list): The list of trajectories.
+            x (int/float): A longitude value according to EPSG:4326.
+            y (int/float): A latitude value according to EPSG:4326.
+            start_date (:obj:`str`, optional): The begin of a time interval.
+            end_date (:obj:`str`, optional): The begin of a time interval.
+
+         Returns:
+            list: A trajectory object as a list.
+        """
         ds = self.get_datasource()
 
         for obs in self.observations_properties:
@@ -48,20 +64,19 @@ class ImageCollection(Collection):
                     "x": x,
                     "y": y,
                     "grid": self.grid,
-                    "srid" : self.spatial_ref_system["srid"],
+                    "srid": self.spatial_ref_system["srid"],
                     "start_date": start_date,
                     "end_date": end_date,
                     "time": time
                 }
 
-                imageID = ds.get_trajectory(**args)
+                image_id = ds.get_trajectory(**args)
 
                 obs_info = None
                 class_info = None
 
-                if imageID is not None:
+                if image_id is not None:
 
-                    # TODO verficar se os tipos estão corretos
                     # Get Class information by passing trajectory result
                     obs_info = get_date_from_str(time)
                     obs_info = obs_info.strftime(self.temporal["string_format"])
@@ -71,17 +86,18 @@ class ImageCollection(Collection):
                         class_info = obs["class_property_name"]
 
                     elif self.classification_class.get_type() == "Self":
-                        class_info = imageID
+                        class_info = image_id
 
                     else:
 
                         ds_class = self.classification_class.get_class_ds()
 
-                        class_info = ds_class.get_classe(imageID,
-                                                         self.classification_class.get_value(),
+                        class_info = ds_class.get_classe(image_id,
+                                                         self.classification_class.get_class_property_value(),
                                                          self.classification_class.get_class_property_name(),
-                                                         self.classification_class.get_name(),
-                                                         class_system=self.classification_class.get_class_system())
+                                                         self.classification_class.get_property_name(),
+                                                         class_system=self.classification_class
+                                                         .get_classification_system_name())
 
                     trj = {
                         "collection": self.get_name(),