+++ wlts/controller.py
@@ -1,11 +1,12 @@
 #
 # This file is part of Web Land Trajectory Service.
-# Copyright (C) 2019 INPE.
+# Copyright (C) 2019-2020 INPE.
 #
 # Web Land Trajectory Service is free software; you can redistribute it and/or modify it
 # under the terms of the MIT License; see LICENSE file for more details.
 #
 """Controllers of Web Land Trajectory Service."""
+from flask import abort
 
 from wlts.collections.collection_manager import collection_manager
 
@@ -15,26 +16,32 @@ def describe_collection(collection_name):
     collection = collection_manager.get_collection(collection_name)
 
     if collection is None:
-        return collection
+        abort(404, "Collection Not Found")
     try:
-        data = {
-            "name": collection.name,
-            "description": collection.description,
-            "detail": collection.detail,
-            "collection_type": collection.collection_type(),
-            "resolution_unit": {
-                "unit": collection.get_resolution_unit(),
-                "value": collection.get_resolution_value()
-            },
-            "period": {
-                "start_date": collection.get_start_date(),
-                "end_date": collection.get_end_date()
-            },
-            "spatial_extent": collection.get_spatial_extent()
+        classification_system = collection.classification_class
+
+        describe = dict()
+
+        describe["classification_system"] = {
+            "type": classification_system.get_type(),
+            "classification_system_name": classification_system.get_classification_system_name(),
+            "classification_system_id": classification_system.get_classification_system_id()
         }
+        describe["name"]= collection.name
+        describe["description"] = collection.description
+        describe["detail"] = collection.detail
+        describe["collection_type"] = collection.collection_type()
+        describe["resolution_unit"] = {
+            "unit": collection.get_resolution_unit(),
+            "value": collection.get_resolution_value()
+        }
+        describe["period"] = {
+            "start_date": collection.get_start_date(),
+            "end_date": collection.get_end_date()
+        }
+        describe["spatial_extent"] = collection.get_spatial_extent()
 
-        return data
+        return describe
 
-    # TODO melhorar o retorno dos erros
-    except:
-        return None
+    except Exception:
+        abort(403, "Error while retrive colection metadata")