+++ wlts/datasources/wcs.py
@@ -1,31 +1,36 @@
 #
 # This file is part of Web Land Trajectory Service.
-# Copyright (C) 2019 INPE.
+# Copyright (C) 2019-2020 INPE.
 #
 # Web Land Trajectory Service is free software; you can redistribute it and/or modify it
 # under the terms of the MIT License; see LICENSE file for more details.
 #
 """WLTS WCS DataSource."""
+import base64
 import urllib.request
+from functools import lru_cache
 from uuid import uuid4
 from xml.dom import minidom
 
-import base64
 import gdal
 import requests
-from osgeo import osr
 from shapely.geometry import Point
 from werkzeug.exceptions import NotFound
-from functools import lru_cache
 
 from wlts.datasources.datasource import DataSource
-from wlts.utils import get_date_from_str
+from wlts.utils import get_date_from_str, transform_latlong_to_rowcol
+
 
 class WCS:
-    """WCS Utility Class."""
+    """This class implements the WCS client.."""
 
     def __init__(self, host, **kwargs):
-        """Create a WFS client attached to the given host address (an URL)."""
+        """Create a WCS client attached to the given host address (an URL).
+
+        Args:
+            host (str): the server URL.
+            **kwargs: The keyword arguments with credentials to access WFS.
+        """
         invalid_parameters = set(kwargs) - {"auth"}
 
         if invalid_parameters:
@@ -45,7 +50,11 @@ class WCS:
                 self._auth = kwargs['auth']
 
     def _get_image(self, uri):
-        """Get Response."""
+        """Query the WCS service using HTTP GET verb and return the image result.
+
+        Args:
+            uri (str): URL for the WCS server.
+        """
         if self._auth:
             request = urllib.request.Request(uri)
 
@@ -61,24 +70,27 @@ class WCS:
         try:
             response = urllib.request.urlopen(request, timeout=30)
             if response.info().get('Content-Encoding') == 'gzip':
-               return None
+                return None
             else:
                 return response
         except urllib.request.URLError:
             return None
 
     def _get(self, uri):
-        """Get URI."""
+        """Query the WCS service using HTTP GET verb.
+
+        Args:
+            uri (str): URL for the WCS server.
+        """
         response = requests.get(uri, auth=self._auth)
 
-        if (response.status_code) != 200:
+        if response.status_code != 200:
             raise Exception("Request Fail: {} ".format(response.status_code))
 
         return response.content.decode('utf-8')
 
-
     def list_image(self):
-        """List collection."""
+        """Returns the list of all available image in service."""
         url = "{}/{}&request=GetCapabilities&outputFormat=application/json".format(self.host, self.base_path)
 
         doc = self._get(url)
@@ -95,31 +107,24 @@ class WCS:
         return avaliables
 
     def check_image(self, ft_name):
-        """Utility to check image existence in wcs."""
+        """Utility to check image existence in wcs.
+
+        Args:
+            ft_name (str): The image name.
+        """
         images = self.list_image()
 
         if ft_name not in images:
             raise NotFound('Image "{}" not found'.format(ft_name))
 
-
-    def transform_latlong_to_rowcol(self, data_set, lat, long):
-        """Transform the pixel location of a geospatial coordinate."""
-        srs = osr.SpatialReference()
-        srs.ImportFromWkt(data_set.GetProjection())
-
-        srs_lat_long = srs.CloneGeogCS()
-        ct = osr.CoordinateTransformation(srs_lat_long, srs)
-        x, y, _ = ct.TransformPoint(long, lat)
-
-        transform = data_set.GetGeoTransform()
-
-        x = int((x - transform[0]) / transform[1])
-        y = int((transform[3] - y) / -transform[5])
-
-        return x, y
-
     def open_image(self, url, long, lat):
-        """Open Image."""
+        """Return the image value for a location.
+
+        Args:
+            url (str): URL for the WCS server.
+            long (int/float): A longitude value according to EPSG:4326.
+            lat (int/float): A latitude value according to EPSG:4326.
+        """
         image_data = self._get_image(url)
 
         if not image_data:
@@ -131,7 +136,7 @@ class WCS:
         gdal_dataset = gdal.Open(mmap_name)
 
         if gdal_dataset is not None:
-            x, y = self.transform_latlong_to_rowcol(gdal_dataset, lat, long)
+            x, y = transform_latlong_to_rowcol(gdal_dataset, lat, long)
 
             intval = gdal_dataset.GetRasterBand(1).ReadAsArray(x, y, 1, 1).astype("int")
 
@@ -149,7 +154,21 @@ class WCS:
 
     @lru_cache()
     def get_image(self, image, srid, min_x, max_x, min_y, max_y, column, row, time, x, y):
-        """Get Image."""
+        """Mount the url for get a image(coverage) from server based on GetCoverage request.
+
+        Args:
+            image (str): The image(coverage) name to retrieve from service.
+            srid (int): The CRS of the image(coverage).
+            min_x (int/float): The min x the extent of the image(coverage)
+            max_x (int/float): The max x the extent of the image(coverage)
+            min_y (int/float): The min y the extent of the image(coverage)
+            max_y (int/float): The max y the extent of the image(coverage)
+            column (int): Grid resolution.
+            row (int): Grid resolution.
+            time (str): Time dimension.
+            x (int/float): The location x to retrieve.
+            x (int/float): The location y to retrieve.
+        """
         url = "{}/{}&request=GetCoverage&COVERAGE={}&".format(self.host, self.base_path, image)
 
         url += "CRS=EPSG:4326&RESPONSE_CRS=EPSG:{}&".format(srid)
@@ -158,16 +177,21 @@ class WCS:
 
         url += "&FORMAT=GeoTIFF&WIDTH={}&HEIGHT={}&time={}".format(column, row, time)
 
-        featureID = self.open_image(url, x, y)
+        image_value = self.open_image(url, x, y)
 
-        return featureID
+        return image_value
 
 
 class WCSDataSource(DataSource):
-    """WCSDataSource Class."""
+    """This class implemente a WCSDataSource."""
 
     def __init__(self, id, ds_info):
-        """WCS Init Method."""
+        """Create a WCSDataSource.
+
+        Args:
+            id (str): the datasource identifier.
+            ds_info (dict): A datasource information as a dictionary.
+        """
         super().__init__(id)
 
         if 'user' in ds_info and 'password' in ds_info:
@@ -178,11 +202,15 @@ class WCSDataSource(DataSource):
         self.workspace = ds_info['workspace']
 
     def get_type(self):
-        """Get DataSource type."""
+        """Return the datasource type."""
         return "WCS"
 
     def get_trajectory(self, **kwargs):
-        """Return trajectory for wcs."""
+        """Return a trajectory instance for wcs datasource.
+
+        Args:
+            **kwargs: The keyword arguments.
+        """
         invalid_parameters = set(kwargs) - {"image", "temporal",
                                             "x", "y", "srid",
                                             "grid", "start_date", "end_date", "time"}
@@ -192,7 +220,7 @@ class WCSDataSource(DataSource):
         ts = get_date_from_str(kwargs['time'])
         if kwargs['start_date']:
             start_date = get_date_from_str(kwargs['start_date'])
-            if  ts < start_date:
+            if ts < start_date:
                 return None
         if kwargs['end_date']:
             end_date = get_date_from_str(kwargs['end_date'])
@@ -201,14 +229,11 @@ class WCSDataSource(DataSource):
 
         image_name = self.workspace + ":" + kwargs['image']
 
-        # verifica se a image existe no geoserver
-        # self._wcs.check_image(image_name)
-
         min_x, min_y, max_x, max_y = Point(kwargs['x'], kwargs['y']).buffer(0.002).bounds
 
-        imageID = self._wcs.get_image(image_name, kwargs['srid'],
-                                            min_x , min_y, max_x, max_y,
-                                            (kwargs['grid'])['column'], (kwargs['grid'])['row'],
-                                            kwargs['time'], kwargs['x'], kwargs['y'])
+        image_value = self._wcs.get_image(image_name, kwargs['srid'],
+                                      min_x, min_y, max_x, max_y,
+                                      (kwargs['grid'])['column'], (kwargs['grid'])['row'],
+                                      kwargs['time'], kwargs['x'], kwargs['y'])
 
-        return imageID
+        return image_value