+++ CHANGES.rst
@@ -10,4 +10,26 @@
 Changes
 =======
 
-**Under Development!**
+Version 0.2.0 (2020-12-11)
+--------------------------
+
+
+- First experimental version.
+
+- Support for OGC WFS and OGC WCS datasource.
+
+- Support for vector and raster data.
+
+- Documentation system based on Sphinx.
+
+- Documentation integrated to ``Read the Docs``.
+
+- Package support through Setuptools.
+
+- Installation and use instructions.
+
+- Travis CI support.
+
+- Source code versioning based on `Semantic Versioning 2.0.0 <https://semver.org/>`_.
+
+- License: `MIT <https://raw.githubusercontent.com/brazil-data-cube/bdc-db/b-0.2/LICENSE>`_.